refactor: move `app` folder inside `src`
