+++ src/app/_layout.tsx
@@ -1,5 +1,5 @@
 import { Slot, Stack } from 'expo-router';
-import { AuthProvider } from '../src/contexts/auth-provider';
+import { AuthProvider } from '../contexts/auth-provider';
 
 export default function HomeLayout() {
   return (
