+++ src/app/index.tsx
@@ -5,10 +5,10 @@ import {
   StyleSheet,
   View,
 } from 'react-native';
-import { getTransactionData } from '../src/data/get-transaction-data';
-import { TransactionHistoryItem } from '../src/components/transaction-history-item';
+import { getTransactionData } from '../data/get-transaction-data';
+import { TransactionHistoryItem } from '../components/transaction-history-item';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { useAuth } from '../src/contexts/auth-provider';
+import { useAuth } from '../contexts/auth-provider';
 import { useState } from 'react';
 
 export default function TransactionHistory() {
